Fix variable type for model

+++ internal/app/account/repository.go
@@ -6,11 +6,11 @@ import (
 
 // AccountRepository represent the account's repository
 type Repository interface {
-	FindByID(id int64) (*model.User, error)
+	FindByID(id uint) (*model.User, error)
 	FindByUUID(uuid string) (*model.User, error)
 	FindByEmail(email string) (*model.User, error)
 	FindByUsername(username string) (*model.User, error)
 	Update(user *model.User) error
 	Save(user *model.User) error
-	Delete(id int64) error
+	Delete(id uint) error
 }
