+++ internal/app/model/tag.go
@@ -6,10 +6,10 @@ import (
 
 // Tag ...
 type Tag struct {
-	ID        int64     `json:"id"`
+	ID        uint      `json:"id"`
 	Name      string    `json:"name"`
 	Colour    string    `json:"colour"`
 	CreatedAt time.Time `json:"createdAt"`
 	UpdatedAt time.Time `json:"updatedAt"`
-	UserID    int64
+	UserID    uint
 }
