+++ internal/app/tag/service.go
@@ -6,9 +6,9 @@ import (
 
 // TagService represent the tag's service
 type Service interface {
-	FindByID(id int64) (*dto.TagDTO, error)
+	FindByID(id uint) (*dto.TagDTO, error)
 	FindByName(name string) (*dto.TagDTO, error)
-	Save(tagDTO *dto.TagDTO, userId int64) (*dto.TagDTO, error)
+	Save(tagDTO *dto.TagDTO, userId uint) (*dto.TagDTO, error)
 	Update(tagDTO *dto.TagDTO) (*dto.TagDTO, error)
-	Delete(id int64) error
+	Delete(id uint) error
 }
